ref(query-builder): Reorganize files
