+++ static/app/components/searchQueryBuilder/hooks/useHandleSearch.tsx
