+++ static/app/components/searchQueryBuilder/hooks/useQueryBuilderGrid.tsx
