+++ static/app/components/searchQueryBuilder/hooks/useSelectOnDrag.tsx
