+++ static/app/components/searchQueryBuilder/hooks/useUndoStack.tsx
