+++ static/app/components/searchQueryBuilder/utils.tsx
@@ -1,29 +1,17 @@
-import {useCallback} from 'react';
-import {getFocusableTreeWalker} from '@react-aria/focus';
-import type {ListState} from '@react-stately/list';
-import type {Node} from '@react-types/shared';
-
 import {
   BooleanOperator,
   FilterType,
-  filterTypeConfig,
-  interchangeableFilterOperators,
   type ParseResult,
   type ParseResultToken,
   parseSearch,
   type SearchConfig,
-  type TermOperator,
   Token,
   type TokenResult,
 } from 'sentry/components/searchSyntax/parser';
-import {t} from 'sentry/locale';
-import type {Tag, TagCollection} from 'sentry/types';
-import {escapeDoubleQuotes} from 'sentry/utils';
-import {FieldKey, FieldValueType, getFieldDefinition} from 'sentry/utils/fields';
+import type {TagCollection} from 'sentry/types/group';
+import {FieldValueType, getFieldDefinition} from 'sentry/utils/fields';
 
 export const INTERFACE_TYPE_LOCALSTORAGE_KEY = 'search-query-builder-interface';
-
-const SHOULD_ESCAPE_REGEX = /[\s"()]/;
 
 function getSearchConfigFromKeys(keys: TagCollection): Partial<SearchConfig> {
   const config = {
@@ -100,15 +88,11 @@ const isSimpleTextToken = (
   return [Token.FREE_TEXT, Token.SPACES].includes(token.type);
 };
 
-export function getKeyLabel(key: Tag) {
-  return key.alias ?? key.key;
-}
-
 /**
  * Collapse adjacent FREE_TEXT and SPACES tokens into a single token.
  * This is useful for rendering the minimum number of inputs in the UI.
  */
-export function collapseTextTokens(tokens: ParseResult | null) {
+function collapseTextTokens(tokens: ParseResult | null) {
   if (!tokens) {
     return null;
   }
@@ -148,114 +132,6 @@ export function tokenIsInvalid(token: TokenResult<Token>) {
   }
 
   return Boolean(token.invalid);
-}
-
-export function getValidOpsForFilter(
-  filterToken: TokenResult<Token.FILTER>
-): readonly TermOperator[] {
-  // If the token is invalid we want to use the possible expected types as our filter type
-  const validTypes = filterToken.invalid?.expectedType ?? [filterToken.filter];
-
-  // Determine any interchangeable filter types for our valid types
-  const interchangeableTypes = validTypes.map(
-    type => interchangeableFilterOperators[type] ?? []
-  );
-
-  // Combine all types
-  const allValidTypes = [...new Set([...validTypes, ...interchangeableTypes.flat()])];
-
-  // Find all valid operations
-  const validOps = new Set<TermOperator>(
-    allValidTypes.flatMap(type => filterTypeConfig[type].validOps)
-  );
-
-  return [...validOps];
-}
-
-export function escapeTagValue(value: string): string {
-  if (!value) {
-    return '';
-  }
-
-  // Wrap in quotes if there is a space or parens
-  return SHOULD_ESCAPE_REGEX.test(value) ? `"${escapeDoubleQuotes(value)}"` : value;
-}
-
-export function unescapeTagValue(value: string): string {
-  return value.replace(/\\"/g, '"');
-}
-
-export function formatFilterValue(token: TokenResult<Token.FILTER>['value']): string {
-  switch (token.type) {
-    case Token.VALUE_TEXT: {
-      if (!token.value) {
-        return token.text;
-      }
-
-      return token.quoted ? unescapeTagValue(token.value) : token.text;
-    }
-    case Token.VALUE_RELATIVE_DATE:
-      return t('%s', `${token.value}${token.unit} ago`);
-    default:
-      return token.text;
-  }
-}
-
-export function getDefaultFilterValue({key}: {key: string}): string {
-  const fieldDef = getFieldDefinition(key);
-
-  if (!fieldDef) {
-    return '""';
-  }
-
-  if (key === FieldKey.IS) {
-    return 'unresolved';
-  }
-
-  switch (fieldDef.valueType) {
-    case FieldValueType.BOOLEAN:
-      return 'true';
-    case FieldValueType.INTEGER:
-    case FieldValueType.NUMBER:
-      return '100';
-    case FieldValueType.DATE:
-      return '-24h';
-    case FieldValueType.STRING:
-    default:
-      return '""';
-  }
-}
-
-export function shiftFocusToChild(
-  element: HTMLElement,
-  item: Node<ParseResultToken>,
-  state: ListState<ParseResultToken>
-) {
-  // Ensure that the state is updated correctly
-  state.selectionManager.setFocusedKey(item.key);
-
-  // When this row gains focus, immediately shift focus to the input
-  const walker = getFocusableTreeWalker(element);
-  const nextNode = walker.nextNode();
-  if (nextNode) {
-    (nextNode as HTMLElement).focus();
-  }
-}
-
-export function useShiftFocusToChild(
-  item: Node<ParseResultToken>,
-  state: ListState<ParseResultToken>
-) {
-  const onFocus = useCallback(
-    (e: React.FocusEvent<HTMLDivElement, Element>) => {
-      shiftFocusToChild(e.currentTarget, item, state);
-    },
-    [item, state]
-  );
-
-  return {
-    shiftFocusProps: {onFocus},
-  };
 }
 
 export function isDateToken(token: TokenResult<Token.FILTER>) {
